+++ src/main/java/org/example/board_project/mapper/BoardMapper.java
@@ -1,8 +1,8 @@
 package org.example.board_project.mapper;
 
 import org.apache.ibatis.annotations.Mapper;
-import org.example.board_project.model.dto.requestDTO.BoardListRequestDTO;
-import org.example.board_project.model.dto.requestDTO.WriteBoardRequestDTO;
+import org.example.board_project.model.dto.requestDTO.board.BoardListRequestDTO;
+import org.example.board_project.model.dto.requestDTO.board.WriteBoardRequestDTO;
 import org.example.board_project.model.entity.Board;
 
 import java.util.List;
